signal: Add support properly unblocking signals on removal
If l_signal_remove() is called it must reverse the effects from
l_signal_create(). It must use SIG_UNBLOCK and not the old mask, as
another call of l_signal_create() may have changed it and using
SIG_SETMASK with the oldmask would not work.

If the signal is being masked multiple times, account that and do not
unblock it until the last user is done.