dbus: Optimize internal storage
We now store the parameter signature first, which is backward from the
ordering in which the parameters are specified by the caller (this
ordering is intended to be the same as the natural order in C syntax).

By storing the parameter signature first, we can very quickly find the
parameter signature in order to compare it to the signature of incoming
method calls.