checksum: Add l_checksum_clone API
In the kernel API we use and also in GLib's API the l_checksum_get_digest()
equivalents can only be called once on each object meaning that in a
situation when multiple digest values at different points of a
running stream are needed the API becomes impractical.  The new _clone()
call returns a clone of the checksum's state at the current time, on
which get_digest() can be called.