cipher: Add l_asymmetric_cipher & RSA support
RSA support is required for different X509 digital certificate and TLS
operations and is since recently present in the kernel.  Being an
asymetric key cipher the API is not the same as that used by the
current symmetric ciphers in cipher.c based on the algif_skcipher driver.
At this time there's no corresponding algif_akcipher for userspace to
asymmetric key ciphers in Linux mainline or the crypto tree so this code
works against a patch that applies on the mainline kernel.
Another algif_akcipher implementation has been posted this week to the
linux-kernel-cryptapi list but comments so far indicate that it will be a
while before it is finalised so temporarily acceping this implementation
against a patched kernel may be a better option than waiting for the
mainline support.

It is not clear if the kernel is going to include the padding scheme
transformations.  We implement the PKCS#1 v1.5 padding scheme in cipher.c
here.  Raw RSA primitives are generally not used without a padding
scheme so we don't expose access to the RSA without PKCS#1 v1.5 padding.