cipher: Fix valgrind issues
Fix a bad size calculation in the RSA key utility and zero the memory
passed to sendmsg() in cipher_operate and l_cipher_set_iv.  Both the
msg and the c_msg_buf are shown as uninitialised by valgrind and one
of them did actually make cipher operations randomly fail.