tls: Keep multiple types of hashes of handshake messages.
We keep a hash of all the handshake messages sent or received so far and
for TLS 1.2 we needed the hash to be a SHA256, although even in 1.2
having more hash types increases the chances of finding a common one
with the peer.  For 1.0 we need MD5 and SHA1 so with this patch we have
an array of different struct l_checksum instances, one for each of the
three hash types.