dbus: Add _dbus_message_builder_mark
This method can mark a checkpoint on the builder, so that if subsequent
operations fail, the builder can be returned to a well-known good state.

This is primarily intended for GetProperties() generators, which need to
deal with getters() potentially returning an error and leaving the
builder in an invalid state.