dbus: Update GVariant message header format
In the current GVariant message format since about Jan 2015 there's no
body size field (it is reserved and must be sent as 0) or header size
information or signature.  The message needs to be parsed as a single
GVariant blob if we want to split it into a header and a body.