gvariant: Allow empty signature
empty signatures are now allowed in enter_struct_dict_common

gvariant_iter_init_internal now checks if the signature is empty before
calling _gvariant_num_children which would return an error in that case.
The gvariant_iter_init_internal caller is responsible for checking if
the signature is allowed in the specific context.