dbus: Better search for matching rules in the filter tree
When adding new filter rules, perform a more complete search for
existing rules that already cover the rule being added by removing the
requierement that levels in the tree are ordered by the match type.
This may add overhead to adding new rules in some situations but will
reduce the number of server-side rules being used and the number of
requests.  This covers a scenario where first a watch is added for:

type=signal, path=/

followed by

type=signal, sender=org.test, path=/

I believe there are some more tricky scenarios which are not covered
specifically when more than one child of an existing node matches some
condition of the rule being added.