key: Rename asymmetric key type to L_KEY_RSA
The kernel automatically determines the correct crypto algorithm using
an OID present in each X.509 key, but it's still important to know which
algorithm it is in userspace so the requested cipher can be matched
up.

RSA is the only asymmetric algorithm supported using the keyctl API as
of v4.8. The crypto algorithm for a key can be parsed out of /proc/keys,
but it may soon be accessible through the keyctl API.