dbus: Handle optional properties in PropertiesChanged
Handle property getter returning an error when building the
PropertiesChanged signal instead of returning NULL.  This is important
because the signal may include multiple properties and if one getter
fails other properties should still be included.  One options is to only
error out if all getters fail, another is to add them to the invalidated
properties list of the signal.  The latter seems to be the right thing
to do for optional properties that may appear and disappear.