dbus: Make adjustments for compilation with clang
clang errors if KDBUS_ITEM_SIZE is a function, because it is used in an
array size inside a struct, where clang does not accept dynamic values.

clang warns about va_lists with elements of type char, since they are
subject to auto-expansion and will be converted to ints either way.