log: Fix systemd journal PRIORITY field logging
When logging messages to the systemd journal the trailing line feed
character was truncated since the message length was incorrectly
reduced by one.  That resulted in the next systemd journal field, in
this case PRIORITY, being written as part of the logged message string
(e.g. "fooPRIORITY=6") rather than as a separate journal field.