unit: Make count_fds() work in fakeroot
fakeroot opens up a TCP socket in a separate thread / behind the scenes.
This is triggered somewhat randomly, so can cause the unit test to fail
sometimes and not other times.

Change the implementation so O_RDWR files are ignored (sockets are
reported by fcntl with such permissions)  This is safe since we always
open file descriptors for unit tests using O_RDONLY mode.