tls: Flush rx buffer on failure
When we send an Alert message to the peer (and our alerts are always
fatal) ignore any buffered messages we've received.  When the
certificate chain verification fails on a client we are usually
processing the third message in a sequence of five received from the
server.  We would then send an Alert and forget the negotiated state,
then process the two remaining messages which would result in two more
Alerts being sent, with the default TLS 1.0 version header.  There is
no need to process those remaining messages.