dbus: Emit InterfacesAdded on new ObjectManager addition
When an Object Manager interface is added to the tree emit
InterfaceAdded signals for all the now-managed objects that may have
existed already.

This way clients that register the Object Manager in the dbus ready
callback will have no issues with InterfacesAdded signals missing if
other interfaces have been registered and added before the callback.