tls: prf: Make 0-length secret handling explicit
It is possible not to use a secret when invoking the TLS prf functions.
Instead of using NULL for the secret, give a hint that it is empty by
passing an empty string.

Additionally, the secret offset logic is changed to explicitly not
perform offset calculations on secret if the secret length is 0.

This change is purely for readability, no functional changes.