Re-enable -Wmissing-field-initializers warning.
The L_PLUGIN_DEFINE() macro was the sole part of ELL that triggered
the missing field initializer warning.  Now that it no longer triggers
that warning drop the -Wno-missing-field-initializers option from the
maintainer mode CFLAGS so that exposure of such problems to the user
can be prevented in the future.