signal: Fix signals not being registered properly
The current signal code would generate the sigmask based on the
signal_list queue and set it inside setup_signalfd.  Unfortunately when
l_signal_create calls setup_signalfd, it has failed to add the new
signal descriptor to the signal_list.  So the signalfd system call
generates / uses the wrong signal mask.

This patch simplifies the whole process.  Instead of generating a signal
mask each call to signalfd, we maintain a signal mask that is
initialized to an empty set on the first call to l_signal_create.  That
mask is maintained as signals are added / removed.

An optimization is also introduced to avoid calling signalfd an extra
time when calling l_signal_create for the first time.