cert: Add cert.c for certificate types and routines
Create a new file for all the generic X.509 certificate utilities so far
living in tls.c and make l_cert and l_certchain public objects.  In this
patch we only add the contructors and destructors.  Since there's not
going to be any way to add a certificate to multiple certchains we add
the prev/next pointers for the doubly-linked list implementation
directly to the l_cert structure.  The private certchain constructor is
going to be used by pem file loaders in pem.c and in
tls_parse_certificate_list.