tls: Support loading multiple root CAs
If a file with multiple concatenated certificates is supplied to
set_cecert, load all of them and verify received certificate chains
aginst any of them instead of just the first one.  This is also what
wpa_supplicant does for files supplied with the ca_cert= setting and
it's actually useful.  Some enterprise CAs have multiple root
certificates and provision clients with a file that contains all of them
concatenated and it does happen that they switch from a certificate
chain using one root to a new chain with a different root for their wifi
APs in which case our network config files break while wpa_supplicant
configs keep working.