cert: Change l_certchain_foreach* signatures
Change the "foreach" to "walk" in l_certchain_foreach_from_leaf/_ca
and in l_cert_foreach_cb_t to not imply that the whole chain is always
traversed because we let the callbacks interrupt the iteration by
returning true.  Drop the return values.  Add comments for the two
functions affected.