tls: Tolerate unsolicited extensions in Server Hello
There are apparently servers that include a "Supported Point Formats"
extension in the Server Hello even if the Client Hello has not asked for
one, not conditionted by the presence of the Supported Elliptic Curves
extension.  Don't fail when we see an extra extensions, only add a log
line about it.  It it rare among the servers I tested though.