unit: test-ecdh: add bytes to l_getrandom wrapper
This should have been put into the commit adding the P384 unit test.
Since P384 works with 48 byte integers we need 48 bytes of randomness
in the l_getrandom wrapper. A check was also added so l_getrandom will
fail if too many bytes are requested (e.g. if a larger curve is added).