checksum: Make l_checksum_get_digest more high-level
Check that the recv() in l_checksum_get_digest has received the expected
number of bytes and update the comment to remove the reference to "bytes
read" which implied this is a low-level wrapper for an fd read.  Now the
caller can pass a buffer longer than strictly needed for the hash, in
which case the first digest_len bytes are used and digest_len is
returned to inform the caller of the number of bytes written, or a
shorter buffer to receive the initial part of the digest.  The caller
then only needs to check if the result is not < 0 instead of checking
for short reads.

The normal usage now may be:

uint8_t a[64];
size_t a_len;

a_len = l_checksum_get_digest(checksum, a, sizeof(a));