Revert "key: Change initialization of struct keyctl_dh_params"
This reverts commit 409111695e76e14a6a86b21068e270a8527302a8.

This fixes a build issue with the 4.20 headers. The merged change
for struct keyctl_dh_params is different from what temporarily
showed up in 4.14.70-73 and 4.18.8-10, so there's a "missing braces"
error. At least the public key changes added new #defines in 4.20
so we have something to use for preprocessor logic