time: basic utility for getting the system time
Timeouts are sometimes overkill for long timeouts, or timeouts that
don't require being immediately serviced once expired. For this
reason the 'time' module was added which allows you to get the system
clock time, as well as some basic APIs to check if two time values are
before/after one another, calculating an offset from a start time, and
finding the diff between two time values. This allows an application
to handle its own timeouts lazily with no tie to the main loop.

The APIs were modeled after the kernel APIs in jiffies.h