key: Work around keyctl_dh_params breakage in certain stable kernels
These stable kernels had a change which renamed one of the members of
struct keyctl_dh_params and broke the ELL build:

v4.14.70, v4.14.71, v4.14.72
v4.18.8, v4.18.9, v4.18.10

Later stable releases in the v4.14.x and v4.18.x series reverted to the
previous member name.

This change allows ELL to build using the kernel declaration of the
struct for v4.20 and later, and provides a correct structure declaration
for older kernels.