ecc: add checks for zero, one and order
To mitigate potential attacks, checks were added to the two
scalar initalization APIs l_ecc_scalar{new,new_random}. After
the scalar is created we check that its not zero or one, and
less than the group order. For new_random we also keep the check
that the scalar is less than the group prime, as this is required
in several RFCs/spec for creating random values.

Reported-By: Mathy Vanhoef <Mathy.Vanhoef@nyu.edu>
Suggested-By: Mathy Vanhoef <Mathy.Vanhoef@nyu.edu>