genl: Make sure request_family updates the info cache
Update the cache when l_genl_request_family is called and a family is
found (or autoloaded) successfully.  This allows clients to create a
genl_family object from within this callback and do not need to care
whether the family was previously loaded or not, relying on
l_genl_discover_families & l_genl_add_family_watch.