dhcp: Decrease re-try delay after initial discovery message
Some of the routers require ‘setup’ time between the moment when
connection is established and before they are ready to respond to
the DHCP discovery messages. This used to cause the DHCP clients
to wait a full delay of +-4 seconds for the next attempt to
rediscover DHCP servers. To address this issue, the time delay
between the initial discovery message and its retransmission is
shorten to milliseconds to give enough time to the routers to get
ready and minimize the delay for the clients.